requirements.txt
datetime
mysql-connector-python
pandas
email_validator